Initial version of the validator with DRKey.
The store holds a type able to create MACs for segment and e2e setup requests,
as well as validate them.
For now, the validator only creates the initial MACs at the initiator AS,
and validates these initial MACs at each transit AS. At the moment,
only the generic requests and the segment setup requests are supported;
missing are the e2e setup requests and all responses.