Extend coliquic test.
Have the test also check that QUIC can receive data intended to another
machine, but that was captured in transit.
This is useful for our colibri control plane, where the colibri service
of a transit AS may receive packets destined to the final destination,
but which were forwarded by the BR to this service.